Solve Real-Int problem and finalize version controlling

+++ osdagMainPage.py
@@ -85,11 +85,10 @@ import shutil
 from pathlib import Path
 
 ############################ Pre-Build Database Updation/Creation #################
-sqlpath = Path('ResourceFiles/Database/test_database.sql')
-sqlitepath = Path('ResourceFiles/Database/test_database.sqlite')
+sqlpath = Path('ResourceFiles/Database/Intg_osdag.sql')
+sqlitepath = Path('ResourceFiles/Database/Intg_osdag.sqlite')
 
 if sqlpath.exists():
-    print(sqlitepath.stat().st_mtime, sqlpath.stat().st_mtime, sqlpath.stat().st_mtime - 1,sqlitepath.stat().st_size)
     if not sqlitepath.exists():
         cmd = 'sqlite3 ' + str(sqlitepath) + ' < ' + str(sqlpath)
         os.system(cmd)
@@ -98,14 +97,14 @@ if sqlpath.exists():
 
     elif sqlitepath.stat().st_size == 0 or sqlitepath.stat().st_mtime < sqlpath.stat().st_mtime - 1:
         try:
-            # sqlitenewpath = Path(str(sqlitepath) + '.new')
-            sqlitenewpath = Path('ResourceFiles/Database/test_database_new.sqlite')
+            sqlitenewpath = Path('ResourceFiles/Database/Intg_osdag_new.sqlite')
             cmd = 'sqlite3 ' + str(sqlitenewpath) + ' < ' + str(sqlpath)
             error = os.system(cmd)
-            # if error != 0:
-            #     raise Exception('SQL to SQLite conversion error')
-            # if sqlitenewpath.stat().st_size == 0:
-            #     raise Exception('SQL to SQLite conversion error')
+            print(error)
+            if error != 0:
+                 raise Exception('SQL to SQLite conversion error 1')
+            if sqlitenewpath.stat().st_size == 0:
+                 raise Exception('SQL to SQLite conversion error 2')
             os.remove(sqlitepath)
             sqlitenewpath.rename(sqlitepath)
             sqlpath.touch()
@@ -418,17 +417,18 @@ class OsdagMainWindow(QMainWindow):
 
     def closeEvent(self, event):
         try:
-            sqlitepath = Path('ResourceFiles/Database/test_database.sqlite')
-            sqlpath = Path('ResourceFiles/Database/test_database.sql')
+            sqlitepath = Path('ResourceFiles/Database/Intg_osdag.sqlite')
+            sqlpath = Path('ResourceFiles/Database/Intg_osdag.sql')
+            precisionscript = 'ResourceFiles/Database/precision.awk'
             if sqlitepath.exists() and (
                     not sqlpath.exists() or sqlpath.stat().st_size == 0 or sqlpath.stat().st_mtime < sqlitepath.stat().st_mtime - 1):
-                sqlnewpath = Path('ResourceFiles/Database/test_database_new.sql')
-                cmd = 'sqlite3 ' + str(sqlitepath) + ' .dump > ' + str(sqlnewpath)
+                sqlnewpath = Path('ResourceFiles/Database/Intg_osdag_new.sql')
+                cmd = 'sqlite3 ' + str(sqlitepath) + ' .dump | gawk -f ' + precisionscript + ' > ' + str(sqlnewpath)
                 error = os.system(cmd)
-                # if error != 0:
-                #     raise Exception('SQLite conversion to SQL error')
-                # if sqlnewpath.stat().st_size == 0:
-                #     raise Exception('SQLite conversion to SQL error')
+                if error != 0:
+                     raise Exception('SQLite conversion to SQL error 1')
+                if sqlnewpath.stat().st_size == 0:
+                     raise Exception('SQLite conversion to SQL error 2')
                 os.remove(sqlpath)
                 sqlnewpath.rename(sqlpath)
                 sqlitepath.touch()
